delete file

+++ controllers/admin.js
@@ -1,6 +1,8 @@
 const Product = require("../models/product");
 const { validationResult } = require("express-validator");
 
+const fileHelper = require("../utils/file");
+
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
     pageTitle: "Add Product",
@@ -28,7 +30,7 @@ exports.postAddProduct = (req, res, next) => {
         price: price,
         description: description,
       },
-      errorMessage: 'Attached file is not image.',
+      errorMessage: "Attached file is not image.",
       validationErrors: [],
     });
   }
@@ -153,6 +155,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save().then((result) => {
@@ -187,7 +190,16 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+
+  Product.findById(prodId)
+    .then((p) => {
+      if (!p) {
+        return next(new Error("Product not found!"));
+      }
+
+      fileHelper.deleteFile(p.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then((result) => {
       console.log("DESTROYED PRODUCT!");
       res.redirect("/admin/products");
